+++ plugins/csp-atmospheres/src/models/bruneton/internal/Implementation.hpp
@@ -32,28 +32,28 @@ struct Params {
     /// The outer vector contains entries for each angle of the phase function. The first item
     /// corresponds to 0° (forward scattering), the last item to 180° (back scattering). The inner
     /// vectors contain the intensity values for each wavelength at the specific angle.
-    std::vector<std::vector<double>> mPhase;
+    std::vector<std::vector<float>> mPhase;
 
     /// Beta_sca per wavelength for the altitude where density is 1.0.
-    std::vector<double> mScattering;
+    std::vector<float> mScattering;
 
     /// Beta_abs per wavelength for the altitude where density is 1.0.
-    std::vector<double> mAbsorption;
+    std::vector<float> mAbsorption;
 
     /// Linear function describing the density distribution from bottom to top. The value at a
     /// specific altitude will be multiplied with the Beta_sca and Beta_abs values above.
-    std::vector<double> mDensity;
+    std::vector<float> mDensity;
   };
 
   /// This struct basically corresponds to the Model::Settings::AbsorbingComponent struct, however
   /// it contains the actual values loaded from the CSV files.
   struct AbsorbingComponent {
     /// Beta_abs per wavelength for N_0
-    std::vector<double> mAbsorption;
+    std::vector<float> mAbsorption;
 
     /// Linear function describing the density distribution from bottom to top. The value at a
     /// specific altitude will be multiplied with the Beta_sca and Beta_abs values above.
-    std::vector<double> mDensity;
+    std::vector<float> mDensity;
   };
 
   /// The atmosphere can contain these components. If no absorbing component was configured, all
@@ -64,13 +64,13 @@ struct Params {
 
   /// The wavelength values, in nanometers, and sorted in increasing order, for which the
   /// phase functions and extinction coefficients in the atmosphere components are given.
-  std::vector<double> mWavelengths;
+  std::vector<float> mWavelengths;
 
   /// See the Model class header for an explanation of these properties.
-  double  mSunAngularRadius;
-  double  mBottomRadius;
-  double  mTopRadius;
-  double  mGroundAlbedo;
+  float  mSunAngularRadius;
+  float  mBottomRadius;
+  float  mTopRadius;
+  float  mGroundAlbedo;
   int32_t mSampleCountOpticalDepth;
   int32_t mSampleCountSingleScattering;
   int32_t mSampleCountMultiScattering;
@@ -89,15 +89,15 @@ struct Params {
   /// (for maximum precision, use the smallest Sun zenith angle yielding negligible sky light
   /// radiance values. For instance, for the Earth case, 102 degrees is a good choice for most cases
   /// (120 degrees is necessary for very high exposure values).
-  double mMaxSunZenithAngle;
+  float mMaxSunZenithAngle;
 };
 
 class Implementation {
  public:
   /// If only three wavelengths are used during rendering, these three are used:
-  static constexpr double kLambdaR = 680.0;
-  static constexpr double kLambdaG = 550.0;
-  static constexpr double kLambdaB = 440.0;
+  static constexpr float kLambdaR = 680.0;
+  static constexpr float kLambdaG = 550.0;
+  static constexpr float kLambdaB = 440.0;
 
   /// The constructor of the class takes all parameters which define the attributes of the
   /// atmosphere. It will allocate various GPU resources.
@@ -120,19 +120,19 @@ class Implementation {
  private:
   void precompute(GLuint fbo, GLuint deltaIrradianceTexture, GLuint deltaMoleculesScatteringTexture,
       GLuint deltaAerosolsScatteringTexture, GLuint deltaScatteringDensityTexture,
-      GLuint deltaMultipleScatteringTexture, glm::dvec3 const& lambdas,
+      GLuint deltaMultipleScatteringTexture, glm::vec3 const& lambdas,
       glm::mat3 const& luminanceFromRadiance, bool blend, unsigned int numScatteringOrders);
 
   void updatePhaseFunctionTexture(
       std::vector<Params::ScatteringComponent> const& scatteringComponents,
-      glm::dvec3 const&                               lambdas);
+      glm::vec3 const&                               lambdas);
 
   const Params  mParams;
   const int32_t mScatteringTextureWidth;
   const int32_t mScatteringTextureHeight;
   const int32_t mScatteringTextureDepth;
 
-  std::function<std::string(glm::dvec3 const&)> mGlslHeaderFactory;
+  std::function<std::string(glm::vec3 const&)> mGlslHeaderFactory;
 
   // To optimize resource usage, this texture stores single molecule-scattering plus all
   // multiple-scattering contributions. The single aerosols scattering is stored in an extra
